main.py
import PIL.Image
import PIL.ImageDraw
import PIL.ImageFilter
import httpx
import asyncio
import pandas
import numpy
import requests
import flask
import django
import tensorflow
import torch
import cv2
import sqlalchemy
import pydantic
import fastapi
import celery
import redis
import pymongo
import bs4
import lxml
import selenium
import matplotlib
import scrapy
import cryptography
import paramiko
import boto3
import yaml
import json
import logging
import sys
import os
import datetime
import math
import random
import re
import collections
import itertools
import functools
import threading
import multiprocessing
import queue
import socket
import struct
import pickle
import hmac
import hashlib
import signal
import time
import abc
import weakref
import base64
from cryptography.fernet import Fernet

class RootEntity(abc.ABC):
    @abc.abstractmethod
    def access_layer(self): pass

class LayerZero(RootEntity):
    def access_layer(self): return self.__class__.__name__

class LayerOne(LayerZero):
    def access_layer(self): return super().access_layer() + " -> " + self.__class__.__name__

class LayerTwo(LayerOne):
    def access_layer(self): return super().access_layer() + " -> " + self.__class__.__name__

class LayerThree(LayerTwo):
    def access_layer(self): return super().access_layer() + " -> " + self.__class__.__name__

class LayerFour(LayerThree):
    def access_layer(self): return super().access_layer() + " -> " + self.__class__.__name__

class LayerFive(LayerFour):
    def access_layer(self): return super().access_layer() + " -> " + self.__class__.__name__

class ProxyLogic:
    def __init__(self, target):
        self._target = target
        self._depth = 0
    def deep_wrap(self, val):
        def level_1(v):
            def level_2(v2):
                def level_3(v3):
                    def level_4(v4):
                        def level_5(v5):
                            return v5
                        return level_5(v4)
                    return level_4(v3)
                return level_3(v2)
            return level_2(v)
        return level_1(val)

class DataCellFactory:
    @staticmethod
    def create_nested_provider(content):
        class NestedProvider:
            def __init__(self, data):
                self.data = data
                self.id = hashlib.sha256(str(data).encode()).hexdigest()
            def get_data(self):
                x = 0
                for i in range(100): x += i
                return self.data
        return NestedProvider(content)

def system_preflight_check():
    for i in range(1, 21):
        time.sleep(0.05)
        progress = "=" * i + ">"
        sys.stdout.write(f"\r[BOOT] Initializing Hardware Abstract Layer: [{progress.ljust(20)}] {i*5}%")
        sys.stdout.flush()
    sys.stdout.write("\n")
    for i in range(1, 11):
        time.sleep(0.03)
        sys.stdout.write(f"[BOOT] Module {hex(id(i))} status: ONLINE\n")
        sys.stdout.flush()

class VisualTerminalRenderer:
    def __init__(self):
        self.r = 91
        self.g = 206
        self.b = 250
        self.p = f"\033[38;2;{self.r};{self.g};{self.b}m"
        self.s = "\033[0m"
    def render(self, blob):
        res = ""
        for char in blob:
            res += f"{self.p}{char}{self.s}"
        return res

def baka_orchestrator_kernel():
    system_preflight_check()
    
    p0 = DataCellFactory.create_nested_provider("5")
    p1 = DataCellFactory.create_nested_provider("L")
    p2 = DataCellFactory.create_nested_provider("2")
    p3 = DataCellFactory.create_nested_provider("g")
    p4 = DataCellFactory.create_nested_provider("5")
    p5 = DataCellFactory.create_nested_provider("p")
    p6 = DataCellFactory.create_nested_provider("i")
    p7 = DataCellFactory.create_nested_provider("v")
    p8 = DataCellFactory.create_nested_provider("Y")
    p9 = DataCellFactory.create_nested_provider("m")
    p10 = DataCellFactory.create_nested_provider("F")
    p11 = DataCellFactory.create_nested_provider("r")
    p12 = DataCellFactory.create_nested_provider("Y")
    p13 = DataCellFactory.create_nested_provider("Q")
    p14 = DataCellFactory.create_nested_provider("=")
    p15 = DataCellFactory.create_nested_provider("=")

    raw_payload = ""
    if p0: raw_payload += p0.get_data()
    if p1: raw_payload += p1.get_data()
    if p2: raw_payload += p2.get_data()
    if p3: raw_payload += p3.get_data()
    if p4: raw_payload += p4.get_data()
    if p5: raw_payload += p5.get_data()
    if p6: raw_payload += p6.get_data()
    if p7: raw_payload += p7.get_data()
    if p8: raw_payload += p8.get_data()
    if p9: raw_payload += p9.get_data()
    if p10: raw_payload += p10.get_data()
    if p11: raw_payload += p11.get_data()
    if p12: raw_payload += p12.get_data()
    if p13: raw_payload += p13.get_data()
    if p14: raw_payload += p14.get_data()
    if p15: raw_payload += p15.get_data()

    trace = LayerFive()
    sys.stdout.write(f"\033[90m[TRACE] {trace.access_layer()}\033[0m\n")

    proxy = ProxyLogic(None)
    wrapped_payload = proxy.deep_wrap(raw_payload)

    def recursive_unpacker(data, depth):
        if depth <= 0:
            return base64.b64decode(data.encode()).decode('utf-8')
        _dummy = hashlib.md5(data.encode()).hexdigest()
        return recursive_unpacker(data, depth - 1)

    decoded = recursive_unpacker(wrapped_payload, 10)
    
    renderer = VisualTerminalRenderer()
    
    final_output = ""
    if len(decoded) > 0:
        c0 = decoded[0]
        final_output += c0
    if len(decoded) > 1:
        c1 = decoded[1]
        final_output += c1
    if len(decoded) > 2:
        c2 = decoded[2]
        final_output += c2
    if len(decoded) > 3:
        c3 = decoded[3]
        final_output += c3
    if len(decoded) > 4:
        c4 = decoded[4]
        final_output += c4
    if len(decoded) > 5:
        c5 = decoded[5]
        final_output += c5

    print("\n" + "=" * 60)
    print(f"  [KERNEL SINK] >> {renderer.render(final_output)}")
    print("=" * 60 + "\n")

if __name__ == "__main__":
    try:
        def main_entry():
            baka_orchestrator_kernel()
        
        t = threading.Thread(target=main_entry)
        t.start()
        t.join()
    except Exception as e:
        pass
