requirements.txt
cryptography>=41.0.7
Pillow>=10.1.0
httpx>=0.25.2
pandas>=2.1.4
numpy>=1.26.2
requests>=2.31.0
flask>=3.0.0
django>=5.0.0
tensorflow>=2.15.0
torch>=2.1.1
opencv-python>=4.8.1
sqlalchemy>=2.0.23
pydantic>=2.5.2
fastapi>=0.104.1
celery>=5.3.6
redis>=5.0.1
pymongo>=4.6.1
beautifulsoup4>=4.12.2
lxml>=4.9.3
selenium>=4.15.2
matplotlib>=3.8.2
scrapy>=2.11.0
paramiko>=3.4.0
boto3>=1.33.0
pyyaml>=6.0.1
scikit-learn>=1.3.2
scipy>=1.11.4
dask[complete]>=2023.11.0
pyarrow>=14.0.1
aiohttp>=3.9.1
networkx>=3.2.1
statsmodels>=0.14.0
plotly>=5.18.0
nltk>=3.8.1
spacy>=3.7.2
jax>=0.4.20
jaxlib>=0.4.20
kivy>=2.2.1
pyqt5>=5.15.10
twisted>=22.10.0
xgboost>=2.0.2
lightgbm>=4.1.0
mxnet>=1.9.1
apache-airflow>=2.8.0
ray>=2.8.1
scikit-image>=0.22.0
vowpalwabbit>=9.9.0
catboost>=1.2.2
onnxruntime>=1.16.3
h5py>=3.10.0
tables>=3.9.1
xlrd>=2.0.1
openpyxl>=3.1.2
colorama>=0.4.6
tqdm>=4.66.0
joblib>=1.3.2
mne>=1.6.0
astropy>=6.0.0
biopython>=1.81
pyopengl>=3.1.7
pyopencl>=2023.1.4
arcade>=2.6.17
Pygments>=2.17.2
scikit-video>=1.1.11
gym>=0.26.2
stable-baselines3>=2.2.1
